requirements.txt
Django==4.2
django-cors-headers==3.14.0
djangorestframework==3.14.0
mysqlclient==2.2.0